[REFACTOR] separate uploadInvoice tests for invoice controller
